+++ packages/fabric/editor-core/examples/5-full-page.tsx
@@ -132,12 +132,11 @@ export default class Example extends React.Component<Props, State> {
               allowCodeBlocks={true}
               allowLists={true}
               allowTextColor={true}
-              allowTables={true}
+              allowTables={{ allowColumnResizing: true }}
               allowJiraIssue={true}
               allowUnsupportedContent={true}
               allowPanel={true}
               allowExtension={true}
-              allowPlaceholderCursor={true}
               allowRule={true}
               allowDate={true}
               {...providers}