+++ packages/fabric/editor-test-helpers/src/schema.ts
@@ -58,6 +58,7 @@ export default createSchema({
     'subsup',
     'emojiQuery',
     'textColor',
+    'confluenceInlineComment',
   ],
   customNodeSpecs: {
     plain: { ...paragraph, content: 'text*', marks: '' },