+++ packages/fabric/media-viewer/src/domain/media-file-attributes.ts
@@ -49,13 +49,19 @@ export class MediaFileAttributesFactory {
 
     const paramsSeparator = resource.indexOf('?') > -1 ? '&' : '?';
 
+    // If the processing status is empty means the observable resolved with error
+    const isError = !details.processingStatus;
+    const processed = details.processingStatus === 'succeeded';
+    const mediaType = artifactFormat ? artifactFormat.type : details.mediaType;
+    const type = isError ? 'error' : (processed ? mediaType : 'non-supported');
+
     return {
       id,
       src: `${serviceHost}${resource}${paramsSeparator}${stringify(
         additionalParams,
       )}`,
       srcDownload: `${serviceHost}${binary}?dl=1`,
-      type: artifactFormat ? artifactFormat.type : details.mediaType,
+      type,
       title: details.name,
       src_hd: video1280 && `${serviceHost}${video1280}`,
       poster: poster1280