+++ packages/utils/docs/src/Props/kindToString.js
@@ -121,7 +121,11 @@ converters.variable = type => {
 };
 
 converters.templateExpression = ({ tag }) => {
-  return `${convert(tag.object)} ${convert(tag.property)}`;
+  return `${convert(tag)}`;
+};
+
+converters.call = ({ callee }) => {
+  return `${convert(callee)}`;
 };
 
 converters.FunctionDeclaration = type => {